Switch to calico fork of logrus
This is to fix a logging issue after bumping logrus to 1.0.6. Packr2
used by felix needed logrus with the `SetOutput` method, but the
fallout was that the messages now often have `<nil> <nil>` instead of
meaningful filename and line number. This only happens when used the
formatting version of log functions (like `Infof`).

The fork of logrus is based on `v1.0.4` with an extra commit adding
the `SetOutput` method to `Logger`.