add NodeAddressKubeInternalIP and NodeAddressKubeExternalIP
Include the Internal/External IP of a node k8s node in the list of
addresses in a calico node.

If an address is already included from BPG as NodeAddressBGPIP, it will
appear twice on the list. It is the users task to filter out the address
of the type it wants and and watch for possible duplicates.