Add go routine stack trace dumping tooling
This commit adds tooling that can be used to dump go routine stack trace information on running go programs. This tooling is useful because it's sometimes difficult to reproduce issues that occur in production, but by enabling this in a production cluster you can get a stack trace of what the program is doing, which will aid in diagnosing possible issues.